+++ packages/commerce-sdk-react/src/hooks/useQuery.ts
@@ -23,6 +23,7 @@ import useConfig from './useConfig'
 import {hasAllKeys} from './utils'
 import {onClient} from '../utils'
 import {CommerceApiProviderProps} from '../provider'
+import {handleInvalidToken, generateCustomEndpointOptions} from './helpers'
 
 /**
  * Helper for query hooks, contains most of the logic in order to keep individual hooks small.
@@ -88,8 +89,8 @@ export const useCustomQuery = (
     queryOptions?: UseQueryOptions<unknown, unknown, unknown, any>
 ) => {
     const config = useConfig()
-    const auth = useAuthContext()
     const logger = config.logger || console
+    const auth = useAuthContext()
     const callCustomEndpointWithAuth = (options: OptionalCustomEndpointClientConfig) => {
         return async () => {
             const {access_token} = await auth.ready()
@@ -100,28 +101,15 @@ export const useCustomQuery = (
             )
 
             return await helpers.callCustomEndpoint(customEndpointOptions).catch(async (error) => {
-                if (error?.response?.status == 401) {
-                    const response = await error?.response?.json()
-                    if (
-                        response?.detail === 'Customer credentials changed after token was issued.'
-                    ) {
-                        logger.info('Login was invalidated. Clearing login state.')
-                        await auth.logout()
+                const {access_token} = await handleInvalidToken(error, auth, logger)
 
-                        // Retry again after resetting auth state
-                        const {access_token} = await auth.ready()
-                        const customEndpointOptions = generateCustomEndpointOptions(
-                            options,
-                            config,
-                            access_token
-                        )
-                        return await helpers.callCustomEndpoint(customEndpointOptions)
-                    } else {
-                        throw error
-                    }
-                } else {
-                    throw error
-                }
+                // Retry again after resetting auth state
+                const customEndpointOptions = generateCustomEndpointOptions(
+                    options,
+                    config,
+                    access_token
+                )
+                return await helpers.callCustomEndpoint(customEndpointOptions)
             })
         }
     }
@@ -149,36 +137,4 @@ export const useCustomQuery = (
     ]
 
     return useReactQuery(queryKey, callCustomEndpointWithAuth(apiOptions), queryOptions)
-}
-
-const generateCustomEndpointOptions = (
-    options: OptionalCustomEndpointClientConfig,
-    config: Omit<CommerceApiProviderProps, 'children'>,
-    access_token: string
-) => {
-    const clientConfig = options.clientConfig || {}
-    const clientHeaders = config.headers || {}
-    return {
-        ...options,
-        options: {
-            method: options.options?.method || 'GET',
-            headers: {
-                Authorization: `Bearer ${access_token}`,
-                ...clientHeaders,
-                ...options.options?.headers
-            },
-            ...options.options
-        },
-        clientConfig: {
-            parameters: {
-                clientId: config.clientId,
-                siteId: config.siteId,
-                organizationId: config.organizationId,
-                shortCode: config.organizationId
-            },
-            proxy: config.proxy,
-            ...clientConfig,
-            throwOnBadResponse: true
-        }
-    }
 }
